+++ code/algorithm-demo/src/com/noob/algorithm/base/template/Main.java
@@ -1,30 +1,43 @@
 package com.noob.algorithm.base.template;
 
+
+import java.util.ArrayList;
+import java.util.List;
+
 public class Main {
 
     // 初始化测试用例
-    public static void initCase(){
+    public static List<int[]> initCase() {
+        List<int[]> caseList = new ArrayList<>();
+
+        // todo 补充测试用例集合
+        caseList.add(new int[]{1, 2, 3});
+
+        return caseList;
+    }
+
+    public static void testSolution1(List<int[]> caseList) {
+
+        for (int[] nums : caseList) {
+
+        }
 
     }
 
-    public static void testSolution1() {
 
-
-    }
-
-
-    public static void testSolution2() {
+    public static void testSolution2(List<int[]> caseList) {
 
     }
 
 
     public static void main(String[] args) {
+        List<int[]> caseList = initCase();
 
-        testSolution1();
+        testSolution1(caseList);
 
         System.out.println("-------------------------------");
 
-        testSolution2();
+        testSolution2(caseList);
 
     }
 }
