+++ code/algorithm-demo/src/com/noob/algorithm/base/template/MainTemplate.java
@@ -1,0 +1,34 @@
+package com.noob.algorithm.base.template;
+
+import com.noob.algorithm.leetcode.q031.Solution1;
+
+import java.util.Arrays;
+
+public class MainTemplate {
+
+    // 初始化测试用例
+    public static void initCase(){
+
+    }
+
+    public static void testSolution1() {
+
+
+    }
+
+
+    public static void testSolution2() {
+
+    }
+
+
+    public static void main(String[] args) {
+
+        testSolution1();
+
+        System.out.println("-------------------------------");
+
+        testSolution2();
+
+    }
+}
